Change class Map to be more integrated to Box2D.
Change class Tile in order to comply to new Map interface and to make it
more independent of it. However it is not tested but it compiles but
there are code to be added in order to make it work as intended.

+++ auxiliary.cpp
@@ -6,10 +6,10 @@ namespace {
 
 	std::random_device rd;
 	std::default_random_engine generator(rd());
-	std::uniform_real_distribution<double> distribution(0.0,1.0);
+	std::uniform_real_distribution<float> distribution(0.0,1.0);
 
 }
 
-double random() {
+float random() {
 	return distribution(generator);
 }
